requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langgraph
pinecone
pymongo
google-api-python-client
google-auth
google-generativeai
pytest